Move ConfigLoader to a .mjs file
